+++ craft_application/commands/remote.py
@@ -60,12 +60,12 @@ class RemoteBuild(ExtensibleCommand):
     @override
     def _fill_parser(self, parser: argparse.ArgumentParser) -> None:
         parser.add_argument(
-            "--recover", action="store_true", help="recover an interrupted build"
+            "--recover", action="store_true", help="Recover an interrupted build"
         )
         parser.add_argument(
             "--launchpad-accept-public-upload",
             action="store_true",
-            help="acknowledge that uploaded code will be publicly available.",
+            help="Acknowledge that uploaded code will be publicly available",
         )
         parser.add_argument(
             "--launchpad-timeout",
@@ -182,6 +182,7 @@ class RemoteBuild(ExtensibleCommand):
             building: set[str] = set()
             succeeded: set[str] = set()
             uploading: set[str] = set()
+            pending: set[str] = set()
             not_building: set[str] = set()
             for arch, build_state in states.items():
                 if build_state.is_running:
@@ -190,6 +191,8 @@ class RemoteBuild(ExtensibleCommand):
                     succeeded.add(arch)
                 elif build_state == BuildState.UPLOADING:
                     uploading.add(arch)
+                elif build_state == BuildState.PENDING:
+                    pending.add(arch)
                 else:
                     not_building.add(arch)
             progress_parts: list[str] = []
@@ -201,6 +204,8 @@ class RemoteBuild(ExtensibleCommand):
                 progress_parts.append("Uploading: " + ", ".join(sorted(uploading)))
             if succeeded:
                 progress_parts.append("Succeeded: " + ", ".join(sorted(succeeded)))
+            if pending:
+                progress_parts.append("Pending: " + ", ".join(sorted(pending)))
             emit.progress("; ".join(progress_parts))
 
         emit.progress("Fetching build artifacts...")